Refactor Size to size/Size
