+++ foundry/core/sprites/__init__.py
@@ -1,3 +1,3 @@
-from foundry.core.Size import Size, SizeProtocol
+from foundry.core.size.Size import Size, SizeProtocol
 
 SPRITE_SIZE: SizeProtocol = Size(8, 16)
